requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
pinecone==6.0.2
pandas==2.1.4
python-dotenv==1.0.0
langchain==0.1.8
langchain-openai==0.0.5
langchain-core>=0.1.24
langchain-text-splitters==0.0.1
httpx==0.25.2
pydantic==2.5.3
langgraph==0.0.20
